Refactored Tour component
